refactor: move app catalog into operations
