+++ .github/ISSUE_TEMPLATE/app-submission.yml
@@ -8,7 +8,7 @@ body:
       value: |
         Submit a working app that uses Pollinations. An automated pre-review checks the live app and optional repository, then a maintainer reviews it before publication.
 
-        Read the [submission criteria](https://github.com/pollinations/pollinations/blob/main/apps/README.md#submission-criteria) first.
+        The app must work, visibly use Pollinations, have a clear purpose, credit Pollinations, and avoid spam, scams, or deceptive behavior.
 
   - type: input
     id: app-name
