+++ .github/workflows/apps-publish-submissions.yml
@@ -61,7 +61,7 @@ jobs:
           ISSUE_URL: ${{ github.event.issue.html_url }}
         run: |
           TODAY=$(date -u +%F)
-          APPROVED_DATE="$TODAY" node apps/app-management/ingestion/validate-submission.js > "$RUNNER_TEMP/app-validation.json"
+          APPROVED_DATE="$TODAY" node operations/app-management/ingestion/validate-submission.js > "$RUNNER_TEMP/app-validation.json"
           jq -e '.valid == true' "$RUNNER_TEMP/app-validation.json" >/dev/null
           APP_JSON=$(jq -c .app "$RUNNER_TEMP/app-validation.json")
           {
@@ -87,11 +87,10 @@ jobs:
         run: |
           BRANCH="auto/app-${ISSUE_NUMBER}-${GITHUB_RUN_ID}-${GITHUB_RUN_ATTEMPT}"
           git switch -c "$BRANCH" origin/main
-          NEW_APP="$APP_JSON" node apps/app-management/catalog.js prepend
-          node apps/app-management/generate-catalog-outputs.js
+          NEW_APP="$APP_JSON" node operations/app-management/app.js prepend
           git config user.name "$BOT_LOGIN"
           git config user.email "${BOT_ID}+${BOT_LOGIN}@users.noreply.github.com"
-          git add apps/catalog.json apps/GREENHOUSE.md README.md
+          git add operations/app-management/app.json
           git commit -m "Add ${APP_NAME} to ${APP_CATEGORY}
 
           Co-authored-by: ${ISSUE_AUTHOR} <${USER_ID}+${ISSUE_AUTHOR}@users.noreply.github.com>
