+++ .github/workflows/data-sync-app-catalog-tinybird.yml
@@ -3,7 +3,7 @@ name: Data / Sync app catalog to Tinybird
 on:
   push:
     branches: [main]
-    paths: [apps/catalog.json]
+    paths: [operations/app-management/app.json]
   schedule:
     # Run at 05:00 UTC daily (after Apps / Update metrics at 04:00)
     - cron: '0 5 * * *'
@@ -18,7 +18,7 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v4
 
-      - name: Sync catalog.json to Tinybird
+      - name: Sync app.json to Tinybird
         env:
           TINYBIRD_SYNC_TOKEN: ${{ secrets.TINYBIRD_SYNC_TOKEN }}
-        run: node apps/app-management/sync-catalog-to-tinybird.js
+        run: node operations/app-management/sync-to-tinybird.js
