+++ CONTRIBUTING.md
@@ -27,7 +27,7 @@ We are in profound need of members who can help enhance our documentation. If yo
 
 ### 4. Build Apps with pollinations.ai
 
-Want to build something cool? Check out our **[apps folder](./apps/README.md)**!
+Want to build something cool? Browse the [community apps](https://pollinations.ai/apps) or [submit your own](https://github.com/pollinations/pollinations/issues/new?template=app-submission.yml).
 
 **How it works:**
 
@@ -43,7 +43,7 @@ Create apps using pollinations.ai APIs and get them featured. Each app lives in 
 - Fun projects (games, quizzes)
 - Developer tools (API playgrounds, dashboards)
 
-See the [apps README](./apps/README.md) for detailed guidelines and templates.
+Use the repository's app submission issue form when your app is ready for review.
 
 ### 5. Code Contributions
