--- apps/_templates/html/README.md
@@ -1,39 +1,0 @@
-# App Name
-
-> One sentence about what this does
-
-**💡 Tip:** Before building, [propose your app idea](https://github.com/pollinations/pollinations/issues/new) using the "SubmitApp" template to get early feedback!
-
-## What It Does
-
-Brief explanation (2-3 sentences). What problem does it solve?
-
-## How to Use
-
-1. Open `index.html` in your browser
-2. That's it!
-
-Or visit the live demo: [link here]
-
-## Features
-
-- Feature 1
-- Feature 2
-- Feature 3
-
-## How It Works
-
-Uses pollinations.ai API to [explain what you're doing with it].
-
-```javascript
-// Example API call from your code
-fetch("https://gen.pollinations.ai/image/cool-cat");
-```
-
-## Screenshots
-
-![Demo](./demo.gif)
-
----
-
-Made with pollinations.ai 🐝
