--- apps/_templates/react/README.md
@@ -1,47 +1,0 @@
-# Your App Name
-
-> One sentence about what this does
-
-**💡 Tip:** Before building, [propose your app idea](https://github.com/pollinations/pollinations/issues/new) using the "SubmitApp" template to get early feedback!
-
-## What It Does
-
-Brief explanation (2-3 sentences). What problem does it solve?
-
-## Quick Start
-
-```bash
-cd apps/your-app-name
-npm install
-npm run dev
-```
-
-Open `http://localhost:5173`
-
-## Build for Production
-
-```bash
-npm run build
-```
-
-The `dist/` folder will contain your production-ready app.
-
-## Features
-
--   Feature 1
--   Feature 2
--   Feature 3
-
-## How It Works
-
-Uses Pollinations API to [explain what you're doing].
-
-## Technologies Used
-
--   React
--   Vite
--   Pollinations AI
-
----
-
-Made with Pollinations AI 🐝
