--- apps/app-management/generate-catalog-outputs.js
@@ -1,240 +1,0 @@
-#!/usr/bin/env node
-
-/**
- * Generate apps/GREENHOUSE.md — a curated highlight reel from all categories.
- *
- * Reads apps/catalog.json (source of truth) and writes:
- *   apps/GREENHOUSE.md  – top apps per category + label highlights
- *   README.md            – updates "Recent Apps" section in repo root
- *
- * Usage: node apps/app-management/generate-catalog-outputs.js
- */
-
-const fs = require("node:fs");
-const { CATALOG_FILE, readApps } = require("./catalog.js");
-
-// ── Config ──────────────────────────────────────────────────────────────────
-
-const GARDEN_FILE = "apps/GREENHOUSE.md";
-const ROOT_README = "README.md";
-
-const THIRTY_DAYS_MS = 30 * 86400000;
-const TOP_N = 10;
-
-const CATEGORIES = [
-    { id: "image", label: "Image", emoji: "🖼️" },
-    { id: "video_audio", label: "Video & Audio", emoji: "🎬" },
-    { id: "writing", label: "Write", emoji: "✍️" },
-    { id: "chat", label: "Chat", emoji: "💬" },
-    { id: "games", label: "Play", emoji: "🎮" },
-    { id: "learn", label: "Learn", emoji: "📚" },
-    { id: "bots", label: "Bots", emoji: "🤖" },
-    { id: "build", label: "Build", emoji: "🛠️" },
-    { id: "business", label: "Business", emoji: "💼" },
-];
-
-// ── Parse ───────────────────────────────────────────────────────────────────
-
-function loadApps() {
-    return readApps().map((app) => {
-        return {
-            emoji: app.emoji,
-            name: app.name,
-            url: app.url,
-            description: app.description,
-            category: app.category.toLowerCase(),
-            github: app.githubUsername,
-            repo: app.repositoryUrl,
-            stars: app.repositoryStars || 0,
-            approvedDate: app.approvedDate,
-            byop: app.byop,
-            requests24h: app.requests24h,
-        };
-    });
-}
-
-// ── Classify ────────────────────────────────────────────────────────────────
-
-function isFresh(app) {
-    return (
-        !!app.approvedDate &&
-        new Date(app.approvedDate) >= new Date(Date.now() - THIRTY_DAYS_MS)
-    );
-}
-function isPollen(app) {
-    return app.byop;
-}
-function isBuzz(app) {
-    return app.requests24h >= 100;
-}
-
-// ── Sort: buzz → pollen → stars → newest ────────────────────────────────────
-
-function sortApps(a, b) {
-    const t = +isBuzz(b) - +isBuzz(a);
-    if (t) return t;
-    if (a.byop !== b.byop) return a.byop ? -1 : 1;
-    const s = (b.stars || 0) - (a.stars || 0);
-    if (s) return s;
-    return (b.approvedDate || "").localeCompare(a.approvedDate || "");
-}
-
-// ── Formatting helpers ──────────────────────────────────────────────────────
-
-function appLink(app) {
-    const href = app.url || app.repo;
-    const label = `${app.emoji} ${app.name}`.trim();
-    return href ? `[${label}](${href})` : label;
-}
-
-function badgeStr(app) {
-    const b = [];
-    if (isBuzz(app)) b.push("🐝");
-    if (isPollen(app)) b.push("🏵️");
-    if (isFresh(app)) b.push("🫧");
-    return b.join(" ");
-}
-
-function authorStr(app) {
-    const gh = app.github?.replace(/^@/, "");
-    return gh ? `[@${gh}](https://github.com/${gh})` : "";
-}
-
-function starsStr(app) {
-    if (!app.stars) return "";
-    if (app.stars >= 1000)
-        return `${(app.stars / 1000).toFixed(1).replace(/\.0$/, "")}k`;
-    return String(app.stars);
-}
-
-function makeTable(apps) {
-    const header =
-        "| App | Description | Author | ⭐ | |\n|-----|-------------|--------|---:|---|";
-    const rows = apps.map(
-        (a) =>
-            `| ${appLink(a)} | ${a.description || ""} | ${authorStr(a)} | ${starsStr(a)} | ${badgeStr(a)} |`,
-    );
-    return header + "\n" + rows.join("\n");
-}
-
-// ── Generate GREENHOUSE.md ──────────────────────────────────────────────────
-
-function generateGarden(apps) {
-    const buzzApps = apps
-        .filter(isBuzz)
-        .sort((a, b) => b.requests24h - a.requests24h);
-    const pollenApps = apps
-        .filter(isPollen)
-        .sort((a, b) => (b.stars || 0) - (a.stars || 0));
-    const freshApps = apps
-        .filter(isFresh)
-        .sort((a, b) =>
-            (b.approvedDate || "").localeCompare(a.approvedDate || ""),
-        );
-
-    let md = `# 🔆 App Greenhouse
-
-> **${apps.length}** community apps powered by [pollinations.ai](https://pollinations.ai)
-
-🐝 **Buzz** = 100+ requests/24h · 🏵️ **Pollen** = Sign in with Pollinations · 🫧 **Fresh** = Added in the last 30 days
-
-🌐 [Browse all apps on pollinations.ai](https://pollinations.ai/apps) · ✏️ [Submit your app](https://github.com/pollinations/pollinations/issues/new?template=app-submission.yml)
-
----
-
-`;
-
-    // Label highlights
-    if (buzzApps.length > 0) {
-        md += `## 🐝 Buzz — Trending Now\n\n`;
-        md += makeTable(buzzApps.slice(0, TOP_N)) + "\n\n";
-    }
-
-    if (pollenApps.length > 0) {
-        md += `## 🏵️ Pollen — Sign in with Pollinations\n\n`;
-        md += `Users pay with their own balance — developers pay $0. [Learn more →](../BRING_YOUR_OWN_POLLEN.md)\n\n`;
-        md += makeTable(pollenApps.slice(0, TOP_N)) + "\n\n";
-    }
-
-    if (freshApps.length > 0) {
-        md += `## 🫧 Fresh — Recently Added\n\n`;
-        md += makeTable(freshApps.slice(0, TOP_N)) + "\n\n";
-    }
-
-    // Category sections
-    for (const cat of CATEGORIES) {
-        const catApps = apps
-            .filter((a) => a.category === cat.id)
-            .sort(sortApps);
-        if (catApps.length === 0) continue;
-
-        md += `---\n\n## ${cat.emoji} ${cat.label}\n\n`;
-        md += makeTable(catApps.slice(0, TOP_N)) + "\n";
-        if (catApps.length > TOP_N) {
-            md += `\n<details><summary>Show all ${catApps.length} apps</summary>\n\n`;
-            md += makeTable(catApps.slice(TOP_N)) + "\n";
-            md += `\n</details>\n`;
-        }
-        md += "\n";
-    }
-
-    return md;
-}
-
-// ── Update root README ──────────────────────────────────────────────────────
-
-function updateRootReadme(apps) {
-    const last10 = apps.slice(0, 10);
-    const rows = last10.map(
-        (a) => `| ${appLink(a)} | ${a.description || ""} | ${authorStr(a)} |`,
-    );
-
-    const section = `## 🆕 Recent Apps
-
-| Name | Description | Author |
-|------|-------------|--------|
-${rows.join("\n")}
-
-[Browse all apps →](apps/GREENHOUSE.md)`;
-
-    if (!fs.existsSync(ROOT_README)) return;
-
-    let readme = fs.readFileSync(ROOT_README, "utf8");
-    const marker = "## 🆕 Recent Apps";
-    const startIdx = readme.indexOf(marker);
-
-    if (startIdx !== -1) {
-        const afterMarker = readme.substring(startIdx + marker.length);
-        const nextSection = afterMarker.search(/\n## /);
-        const endIdx =
-            nextSection !== -1
-                ? startIdx + marker.length + nextSection
-                : readme.length;
-        readme =
-            readme.substring(0, startIdx) + section + readme.substring(endIdx);
-    } else {
-        const firstHeading = readme.search(/\n## /);
-        if (firstHeading !== -1) {
-            readme =
-                readme.substring(0, firstHeading) +
-                "\n\n" +
-                section +
-                readme.substring(firstHeading);
-        }
-    }
-
-    fs.writeFileSync(ROOT_README, readme);
-    console.log("  ✅ Updated README.md (recent apps)");
-}
-
-// ── Main ────────────────────────────────────────────────────────────────────
-
-const apps = loadApps();
-console.log(`Parsed ${apps.length} apps from ${CATALOG_FILE}`);
-
-fs.writeFileSync(GARDEN_FILE, generateGarden(apps));
-console.log(`  ✅ Generated ${GARDEN_FILE}`);
-
-updateRootReadme(apps);
-
-console.log("\n🔆 Greenhouse updated!");
