+++ enter.pollinations.ai/observability/endpoints/app_directory_public.pipe
@@ -1,5 +1,5 @@
 DESCRIPTION >
-    Public community app directory — the apps/catalog.json showcase synced daily into
+    Public community app directory — operations/app-management/app.json synced daily into
     Tinybird (app_directory). Every field here is already published on the
     Pollinations website, so this endpoint is gated by the public read token.
     Returns the full directory; filter or paginate client-side with the
