+++ gen.pollinations.ai/src/docs/public-stats.md
@@ -17,7 +17,7 @@ p.eyJ1IjogImFjYTYzZjc5LThjNTYtNDhlNC05NWJjLWEyYmFjMTY0NmJkMyIsICJpZCI6ICI5ZWZmMG
 | `GET https://gen.pollinations.ai/v1/models/status` | `minutes` (60, max 10080) | Per-model health in a recent window: 2xx/4xx/5xx counts, latency p50/p95. Use this cached gateway instead of calling Tinybird directly. |
 | `GET /v0/pipes/weekly_health_stats.json` | `weeks_back` (12) | Weekly service availability (`2xx / (2xx + 5xx)`, cache excluded) and latency |
 | `GET /v0/pipes/app_top_weekly.json` | — | Top 10 registered apps owned by showcase contributors, by request count over the last 7 days. The owner is listed in the directory; the returned app may be any of their registered apps |
-| `GET /v0/pipes/app_directory_public.json` | `category`, `platform`, `limit` (1000) | The community app directory ([apps/catalog.json](https://github.com/pollinations/pollinations/blob/main/apps/catalog.json)) |
+| `GET /v0/pipes/app_directory_public.json` | `category`, `platform`, `limit` (1000) | The community app directory ([app.json](https://github.com/pollinations/pollinations/blob/main/operations/app-management/app.json)) |
 
 Tinybird responses are JSON: a `data` array of rows plus a `meta` array typing
 each column. Append `&token=<public-read-token>` to authenticate them. The
