+++ operations/app-management/performance/update-metrics.js
@@ -1,13 +1,13 @@
 #!/usr/bin/env node
 
 /**
- * Update app metrics in apps/catalog.json:
+ * Update app metrics in operations/app-management/app.json:
  *   - GitHub stars (from GitHub API)
  *   - BYOP status (from Tinybird — apps using secret API keys)
  *   - Request count in last 24h (from Tinybird — by referrer domain)
  *   - Sort rows: BYOP first → requests desc → stars desc
  *
- * Usage: node apps/app-management/performance/update-metrics.js [options]
+ * Usage: node operations/app-management/performance/update-metrics.js [options]
  *   --dry-run    Show changes without modifying files
  *   --verbose    Show detailed output
  *
@@ -17,7 +17,7 @@
  */
 
 const https = require("node:https");
-const { CATALOG_FILE, readApps, writeApps } = require("../catalog.js");
+const { APP_FILE, readApps, writeApps } = require("../app.js");
 
 const GITHUB_API = "api.github.com";
 const TINYBIRD_HOST = "api.europe-west2.gcp.tinybird.co";
@@ -377,9 +377,7 @@ async function main() {
         }
 
         writeApps(catalog);
-        console.log(
-            `${colors.green}✅ Updated ${CATALOG_FILE}${colors.reset}\n`,
-        );
+        console.log(`${colors.green}✅ Updated ${APP_FILE}${colors.reset}\n`);
     }
 
     // Summary
