+++ pollinations.ai/src/copy/constants.ts
@@ -4,7 +4,7 @@
 export const COPY_CONSTANTS = {
     // External data sources
     appsFilePath:
-        "https://raw.githubusercontent.com/pollinations/pollinations/main/apps/catalog.json",
+        "https://raw.githubusercontent.com/pollinations/pollinations/main/operations/app-management/app.json",
 
     // API
     apiBaseUrl: "gen.pollinations.ai",
